Display quality of signals & fix filters (fixes #9)
This PR implements basic display of signal quality, and fixes filters to
respect quality.

Unfortunately, the modding API does not seem to have support for showing
both qualities and counts on the same icon. So we switch to using labels
with icons instead of buttons, which requires increasing column width.

- [ ] Add a preference to only enable the new display when asked by user